+++ src/api/postsApi.js
@@ -6,6 +6,7 @@
 
 import axios from 'axios'
 import { POSTS_API_URL } from './apis'
+import { isNetworkEnabled, loadLocalData, findLocalDataById } from '../utils/dataSourceManager'
 
 /**
  * Posts API 관련 함수들
@@ -13,38 +14,72 @@ import { POSTS_API_URL } from './apis'
 export const postsApi = {
   // 모든 게시글 조회
   getAll: async () => {
-    const response = await axios.get(POSTS_API_URL)
-    return response.data
+    if (isNetworkEnabled()) {
+      const response = await axios.get(POSTS_API_URL)
+      return response.data
+    } else {
+      return await loadLocalData('posts')
+    }
   },
 
   // 특정 게시글 조회
   getById: async (id) => {
-    const response = await axios.get(`${POSTS_API_URL}/${id}`)
-    return response.data
+    if (isNetworkEnabled()) {
+      const response = await axios.get(`${POSTS_API_URL}/${id}`)
+      return response.data
+    } else {
+      return await findLocalDataById('posts', id)
+    }
   },
 
   // 새 게시글 추가
   create: async (postData) => {
-    const response = await axios.post(POSTS_API_URL, postData)
-    return response.data
+    if (isNetworkEnabled()) {
+      const response = await axios.post(POSTS_API_URL, postData)
+      return response.data
+    } else {
+      const newPost = {
+        id: Date.now(),
+        ...postData
+      }
+      console.log('📄 [Local Mode] Created post:', newPost)
+      return newPost
+    }
   },
 
   // 게시글 수정
   update: async (id, postData) => {
-    const response = await axios.put(`${POSTS_API_URL}/${id}`, postData)
-    return response.data
+    if (isNetworkEnabled()) {
+      const response = await axios.put(`${POSTS_API_URL}/${id}`, postData)
+      return response.data
+    } else {
+      const existingPost = await findLocalDataById('posts', id)
+      const updatedPost = { ...existingPost, ...postData }
+      console.log('✏️ [Local Mode] Updated post:', updatedPost)
+      return updatedPost
+    }
   },
 
   // 게시글 삭제
   remove: async (id) => {
-    const response = await axios.delete(`${POSTS_API_URL}/${id}`)
-    return response.data
+    if (isNetworkEnabled()) {
+      const response = await axios.delete(`${POSTS_API_URL}/${id}`)
+      return response.data
+    } else {
+      console.log('🗑️ [Local Mode] Deleted post with id:', id)
+      return { success: true, id }
+    }
   },
 
   // 다중 게시글 삭제
   deleteMany: async (ids) => {
-    const deletePromises = ids.map(id => postsApi.remove(id))
-    await Promise.all(deletePromises)
-    return ids
+    if (isNetworkEnabled()) {
+      const deletePromises = ids.map(id => postsApi.remove(id))
+      await Promise.all(deletePromises)
+      return ids
+    } else {
+      console.log('🗑️ [Local Mode] Bulk deleted posts with ids:', ids)
+      return ids
+    }
   }
 }
